+++ src/main/scala/chisel3/tester/experimental/package.scala
@@ -19,33 +19,4 @@ import firrtl.ExecutionOptionsManager
 package object experimental {
   type TesterOptions = chisel3.tester.internal.TesterOptions
   val TesterOptions = chisel3.tester.internal.TesterOptions  // expose this internal object, whose "API" is unstable
-
-  implicit class ChiselScalatestOptionBuilder[T <: MultiIOModule](x: ChiselScalatestTester#TestBuilder[T]) {
-    def withExecOptions(opt: firrtl.ExecutionOptionsManager): ChiselScalatestTester#TestBuilder[T] = {
-      new x.outer.TestBuilder[T](x.dutGen, Some(opt), x.testOptions)
-    }
-    def withTesterOptions(opt: TesterOptions): ChiselScalatestTester#TestBuilder[T] = {
-      new x.outer.TestBuilder[T](x.dutGen, x.execOptions, Some(opt))
-    }
-  }
-
-  /** This provides a quick and dirty way to poke clocks. Inter-thread dependency is NOT checked,
-    * so it is up to you to understand the thread ordering semantics to use this correctly.
-    * Note that thread ordering IS deterministic, so you will NOT get a nondeterministic test.
-    */
-  implicit class UncheckedPokeableClock(signal: Clock) {
-    def high(): Unit = {
-      if (DataMirror.directionOf(signal) != ActualDirection.Input) {
-        throw new UnpokeableException("Cannot only poke inputs")
-      }
-      Context().backend.pokeClock(signal, true)
-    }
-
-    def low(): Unit = {
-      if (DataMirror.directionOf(signal) != ActualDirection.Input) {
-        throw new UnpokeableException("Cannot only poke inputs")
-      }
-      Context().backend.pokeClock(signal, false)
-    }
-  }
 }
